+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased][unreleased]
 ### Fixed
+- Unnecessary tax api calls when PayPal Express Checkout used from cart or product page
 - No inventory check when adding an item to the cart that was just ordered
 - Magento admin can place $0 orders regardless of configured payment types.