+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.2.6] - 2015-01-29
+
+### Fixed
+- Magento no longer displays erroneous success message with SUPEE-3345 patch applied.
+
 ## [1.2.5] - 2015-01-23
 
 ### Fixed
@@ -89,6 +94,8 @@ All notable changes to this project will be documented in this file.
 - Gift card PIN is not submitted with the order
 - Product import not importing color descriptions
 
+[1.2.6]: https://github.com/eBayEnterprise/magento-retail-order-management/compare/1.2.5...1.2.6
+[1.2.5]: https://github.com/eBayEnterprise/magento-retail-order-management/compare/1.2.4...1.2.5
 [1.2.4]: https://github.com/eBayEnterprise/magento-retail-order-management/compare/1.2.3...1.2.4
 [1.2.3]: https://github.com/eBayEnterprise/magento-retail-order-management/compare/1.2.2...1.2.3
 [1.2.2]: https://github.com/eBayEnterprise/magento-retail-order-management/compare/1.2.1...1.2.2